+++ python/packages/main/agent_framework/__init__.py
@@ -16,4 +16,4 @@ from ._memory import *  # noqa: F403
 from ._threads import *  # noqa: F403
 from ._tools import *  # noqa: F403
 from ._types import *  # noqa: F403
-from .workflow import *  # noqa: F403
+from ._workflow import *  # noqa: F403
