+++ python/packages/main/agent_framework/_workflow/_magentic.py
